requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.5
PyPDF2==3.0.1
faiss-cpu==1.8.0
sentence-transformers==2.7.0
numpy==1.24.3
